Ref(Program.scala): moved intermediate representations to a new package in analysis
